% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaFinder.R
\name{deltaFinder}
\alias{deltaFinder}
\title{deltaFinder}
\usage{
deltaFinder(a, lambda1, lambda2, gamma)
}
\arguments{
\item{a}{Possbile update value.}

\item{lambda1}{Tuning parameter pertaining to importance of groups.}

\item{lambda2}{Tuning parameter pertaining to importance of individuals.}

\item{gamma}{Scaling / thresholding value.}
}
\value{
L2 Norm minus one of a given delta group defined by
given lambda1, lambda2, and gamma.
}
\description{
Function to find the appropriate group update. A potential
update value is given and then the corresponding sparse group
lasso penalty value is calculated.
}
\details{
Internal function used by bisee.
}
\note{
Internal function.
}
\author{
Gregory Vaughan
}
\keyword{internal}
